# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: UCM API
  version: "1.0.0"
  description: |
    United Care Mobility — Multi-tenant Medical Transportation Management System API.

    Manages fleets, drivers, patients, trips, clinics, billing, and dispatch across multiple cities.

    ## Authentication
    All protected endpoints require a Bearer JWT token obtained via `POST /api/auth/login-jwt`.
    Include the token in the `Authorization` header: `Authorization: Bearer <token>`.

    ## Multi-Tenancy
    The JWT payload includes `companyId` for automatic tenant isolation. All data queries
    are scoped to the authenticated user's company. SUPER_ADMIN users can access data across companies
    by setting the `x-company-id` header.

    ## Roles
    - **SUPER_ADMIN** — Platform-wide access
    - **ADMIN** / **COMPANY_ADMIN** — Company-level administration
    - **DISPATCH** — Trip and driver management
    - **DRIVER** — Driver portal access
    - **VIEWER** — Read-only company access
    - **CLINIC_ADMIN** / **CLINIC_USER** / **CLINIC_VIEWER** — Clinic portal access

    ## Subscription Enforcement
    Companies with `subscription_enabled = true` are subject to subscription status checks
    and quota enforcement. Inactive subscriptions return `SUBSCRIPTION_INACTIVE` (403).
    Quota violations return `QUOTA_EXCEEDED` (403).
  contact:
    name: UCM Engineering
  license:
    name: MIT
servers:
  - url: "{protocol}://{host}"
    description: UCM API Server
    variables:
      protocol:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication token. Obtain via `POST /api/auth/login-jwt`.
        The JWT payload includes `companyId` for tenant scoping.
  schemas:
    WebhookResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        secret:
          type: string
          description: HMAC SHA256 signing secret (only returned on creation)
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    WebhookListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
      required: [message]
...
  responses:
    Unauthorized:
      description: Authentication required — missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: Unauthorized
...
paths:
...
  /api/webhooks:
    post:
      tags: [Webhooks]
      summary: Create webhook endpoint
      description: Register a new webhook endpoint for the authenticated company. Returns the webhook secret (only shown once).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                  description: HTTPS endpoint URL to receive webhook payloads
                events:
                  type: array
                  items:
                    type: string
                    enum:
                      - trip.created
                      - trip.assigned
                      - trip.started
                      - trip.completed
                      - trip.cancelled
                      - driver.location_updated
                      - invoice.created
                      - invoice.paid
                      - subscription.updated
                  description: List of events to subscribe to
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
        "400":
          description: Invalid input
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: COMPANY_ADMIN role required

    get:
      tags: [Webhooks]
      summary: List company webhooks
      description: List all webhook endpoints for the authenticated company. Secrets are not included.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: COMPANY_ADMIN role required
  /api/webhooks/{id}:
    patch:
      tags: [Webhooks]
      summary: Update webhook
      description: Update a webhook endpoint's URL, events, or active status.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                    enum:
                      - trip.created
                      - trip.assigned
                      - trip.started
                      - trip.completed
                      - trip.cancelled
                      - driver.location_updated
                      - invoice.created
                      - invoice.paid
                      - subscription.updated
                active:
                  type: boolean
      responses:
        "200":
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookListItem"
        "400":
          description: Invalid input
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Webhook not found

    delete:
      tags: [Webhooks]
      summary: Deactivate webhook
      description: Soft-delete a webhook by setting active to false.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Webhook deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Webhook not found